Source code documentation

WorldDynamics constants

WorldDynamics functions

Functions in functions.jl

This functions correspond to DYNAMO functions used in the World3 model.

Main.WorldDynamics.clip — Method

clip(f1, f2, va, th)

Return f1 if the value v is greater than the threshold th, f2 otherwise. This function correspond to the CLIP (also called FIFGE) function in the DYNAMO language.

Main.WorldDynamics.interpolate — Method

interpolate(x, y, xs)

Return the value of a function with input x, by linearly interpolating the function itself through the table y and the range xs. If x is out of the range, the value at the corresponding extremity is returned. This function correspond to the TABHL function in the DYNAMO language. This latter function receives a table (that is, y), a value (that is, x), a left and a right extreme of an interval (that is, xs), and an increment value.

Main.WorldDynamics.max — Method

max(v1, v2)

Return the maximum between v1 and v2. This function correspond to the MAX function in the DYNAMO language.

Main.WorldDynamics.min — Method

min(v1, v2)

Return the minimum between v1 and v2. This function correspond to the MIN function in the DYNAMO language.

Main.WorldDynamics.step — Method

step(t, hght, sttm)

Return 0 if the value t is smaller than the threshold sttm, hght otherwise. This function correspond to the STEP function in the DYNAMO language.

Main.WorldDynamics.switch — Method

switch(v1, v2, z)

Return v1 if the value z is approximately 0 with tolerance 1e-16, v2 otherwise. This function correspond to the SWITCH (also called FIFZE) function in the DYNAMO language.

Functions in plotvariables.jl

Functions in solvesystems.jl