Source code documentation

WorldDynamics constants

WorldDynamics functions

Functions in functions.jl

This functions correspond to DYNAMO functions used in the World3 model.

WorldDynamics.clip — Method

clip(f1, f2, va, th)

Return f1 if the value v is greater than the threshold th, f2 otherwise. This function correspond to the CLIP (also called FIFGE) function in the DYNAMO language.

WorldDynamics.interpolate — Method

interpolate(x, y, xs)

Return the value of a function with input x, by linearly interpolating the function itself through the table y and the range xs. If x is out of the range, the value at the corresponding extremity is returned. This function correspond to the TABHL function in the DYNAMO language. This latter function receives a table (that is, y), a value (that is, x), a left and a right extreme of an interval (that is, xs), and an increment value.

WorldDynamics.step — Method

step(t, hght, sttm)

Return 0 if the value t is smaller than the threshold sttm, hght otherwise. This function correspond to the STEP function in the DYNAMO language.

WorldDynamics.switch — Method

switch(v1, v2, z)

Return v1 if the value z is approximately 0 with tolerance 1e-16, v2 otherwise. This function correspond to the SWITCH (also called FIFZE) function in the DYNAMO language.

Functions in plotvariables.jl

WorldDynamics.plotvariables — Method

plotvariables(solution, xrange, variables::Vector{<:NTuple{4, Any}}; title="", showaxis=true, showlegend=true, linetype="lines", colored=true)

Plot the values of the variables in the vector variables obtained by the ODE system solution (normally, obtained by using the solve function in solvesystems.jl) in the specified xrange interval. For each variable, the vector variables includes a quadruple, containing the Julia variable, its range, and its symbolic name to be shown in the plot.

Functions in solvesystems.jl

WorldDynamics.compose — Method

compose(systems::Vector{ODESystem}, connection_eqs::Vector{Equation})

Return the ODE system obtained by composing the ODE systems in the vector systems and by making use of the variable equalities in connection_eqs. Normally, each ODE systems in systems corresponds to a subsystem of a system in the World3 model, and the variable equalities specify which variables are shared between the subsystems.

WorldDynamics.solve — Method

solve(system::ODESystem, timespan; solver=AutoVern9(Rodas5())

Return the solution of the system ODE system in the timespan interval (for the available different ODE system solvers, see the documentation of DifferentialEquations.jl).

We use the AutoVern9(Rodas5()) solver since it is among the suggested ones in the documentation of DifferentialEquations.jl, and among those we tested, it is the one that works best.

Reproducing figures

Agriculture system

WorldDynamics.World3.Agriculture.fig_69a — Method

Reproduce Fig 4.69a. The original figure is presented on Chapter 4.

WorldDynamics.World3.Agriculture.fig_69b — Method

Reproduce Fig 4.69b. The original figure is presented on Chapter 4.

WorldDynamics.World3.Agriculture.fig_69c — Method

Reproduce Fig 4.69c. The original figure is presented on Chapter 4.

WorldDynamics.World3.Agriculture.fig_69d — Method

Reproduce Fig 4.69d. The original figure is presented on Chapter 4.

WorldDynamics.World3.Agriculture.fig_70a — Method

Reproduce Fig 4.70a. The original figure is presented on Chapter 4.

WorldDynamics.World3.Agriculture.fig_70b — Method

Reproduce Fig 4.70b. The original figure is presented on Chapter 4.

WorldDynamics.World3.Agriculture.fig_70c — Method

Reproduce Fig 4.70c. The original figure is presented on Chapter 4.

WorldDynamics.World3.Agriculture.fig_70d — Method

Reproduce Fig 4.70d. The original figure is presented on Chapter 4.

WorldDynamics.World3.Agriculture.fig_72a — Method

Reproduce Fig 4.72a. The original figure is presented on Chapter 4.

WorldDynamics.World3.Agriculture.fig_72b — Method

Reproduce Fig 4.72b. The original figure is presented on Chapter 4.

WorldDynamics.World3.Agriculture.fig_73a — Method

Reproduce Fig 4.73a. The original figure is presented on Chapter 4.

WorldDynamics.World3.Agriculture.fig_73b — Method

Reproduce Fig 4.73b. The original figure is presented on Chapter 4.

WorldDynamics.World3.Agriculture.fig_74a — Method

Reproduce Fig 4.74a. The original figure is presented on Chapter 4.

WorldDynamics.World3.Agriculture.fig_74b — Method

Reproduce Fig 4.74b. The original figure is presented on Chapter 4.

WorldDynamics.World3.Agriculture.fig_75a — Method

Reproduce Fig 4.75a. The original figure is presented on Chapter 4.

WorldDynamics.World3.Agriculture.fig_75b — Method

Reproduce Fig 4.75b. The original figure is presented on Chapter 4.

WorldDynamics.World3.Agriculture.fig_76a — Method

Reproduce Fig 4.76a. The original figure is presented on Chapter 4.

WorldDynamics.World3.Agriculture.fig_76b — Method

Reproduce Fig 4.76b. The original figure is presented on Chapter 4.

WorldDynamics.World3.Agriculture.fig_77a — Method

Reproduce Fig 4.77a. The original figure is presented on Chapter 4.

WorldDynamics.World3.Agriculture.fig_77b — Method

Reproduce Fig 4.77b. The original figure is presented on Chapter 4.

WorldDynamics.World3.Agriculture.fig_78a — Method

Reproduce Fig 4.78a. The original figure is presented on Chapter 4.

WorldDynamics.World3.Agriculture.fig_78b — Method

Reproduce Fig 4.78b. The original figure is presented on Chapter 4.

WorldDynamics.World3.Agriculture.fig_82a — Method

Reproduce Fig 4.82a. The original figure is presented on Chapter 4.

WorldDynamics.World3.Agriculture.fig_82b — Method

Reproduce Fig 4.82b. The original figure is presented on Chapter 4.

WorldDynamics.World3.Agriculture.fig_83a — Method

Reproduce Fig 4.83a. The original figure is presented on Chapter 4.

WorldDynamics.World3.Agriculture.fig_83b — Method

Reproduce Fig 4.83b. The original figure is presented on Chapter 4.

WorldDynamics.World3.Agriculture.fig_84a — Method

Reproduce Fig 4.84a. The original figure is presented on Chapter 4.

WorldDynamics.World3.Agriculture.fig_84b — Method

Reproduce Fig 4.84b. The original figure is presented on Chapter 4.

WorldDynamics.World3.Agriculture.fig_85a — Method

Reproduce Fig 4.85a. The original figure is presented on Chapter 4.

WorldDynamics.World3.Agriculture.fig_85b — Method

Reproduce Fig 4.85b. The original figure is presented on Chapter 4.

WorldDynamics.World3.Agriculture.fig_86a — Method

Reproduce Fig 4.86a. The original figure is presented on Chapter 4.

WorldDynamics.World3.Agriculture.fig_86b — Method

Reproduce Fig 4.86b. The original figure is presented on Chapter 4.

WorldDynamics.World3.Agriculture.fig_87a — Method

Reproduce Fig 4.87a. The original figure is presented on Chapter 4.

WorldDynamics.World3.Agriculture.fig_87b — Method

Reproduce Fig 4.87b. The original figure is presented on Chapter 4.

WorldDynamics.World3.Agriculture.fig_88a — Method

Reproduce Fig 4.88a. The original figure is presented on Chapter 4.

WorldDynamics.World3.Agriculture.fig_88b — Method

Reproduce Fig 4.88b. The original figure is presented on Chapter 4.

WorldDynamics.World3.Agriculture.fig_89a — Method

Reproduce Fig 4.89a. The original figure is presented on Chapter 4.

WorldDynamics.World3.Agriculture.fig_89b — Method

Reproduce Fig 4.89b. The original figure is presented on Chapter 4.

WorldDynamics.World3.Agriculture.fig_90a — Method

Reproduce Fig 4.90a. The original figure is presented on Chapter 4.

WorldDynamics.World3.Agriculture.fig_90b — Method

Reproduce Fig 4.90b. The original figure is presented on Chapter 4.

Capital system

WorldDynamics.World3.Capital.fig_36 — Method

Reproduce Fig 3.36. The original figure is presented on Chapter 3.

WorldDynamics.World3.Capital.fig_37 — Method

Reproduce Fig 3.37. The original figure is presented on Chapter 3.

WorldDynamics.World3.Capital.fig_38 — Method

Reproduce Fig 3.38. The original figure is presented on Chapter 3.

WorldDynamics.World3.Capital.fig_39 — Method

Reproduce Fig 3.39. The original figure is presented on Chapter 3.

WorldDynamics.World3.Capital.fig_40 — Method

Reproduce Fig 3.40. The original figure is presented on Chapter 3.

WorldDynamics.World3.Capital.fig_41 — Method

Reproduce Fig 3.41. The original figure is presented on Chapter 3.

WorldDynamics.World3.Capital.fig_42 — Method

Reproduce Fig 3.42. The original figure is presented on Chapter 3.

WorldDynamics.World3.Capital.fig_43 — Method

Reproduce Fig 3.43. The original figure is presented on Chapter 3.

WorldDynamics.World3.Capital.fig_44 — Method

Reproduce Fig 3.44. The original figure is presented on Chapter 3.

WorldDynamics.World3.Capital.fig_45 — Method

Reproduce Fig 3.45. The original figure is presented on Chapter 3.

WorldDynamics.World3.Capital.fig_46 — Method

Reproduce Fig 3.46. The original figure is presented on Chapter 3.

WorldDynamics.World3.Capital.fig_47 — Method

Reproduce Fig 3.47. The original figure is presented on Chapter 3.

WorldDynamics.World3.Capital.fig_48 — Method

Reproduce Fig 3.48. The original figure is presented on Chapter 3.

Non-Renewable system

WorldDynamics.World3.NonRenewable.fig_25 — Method

Reproduce Fig 5.25. The original figure is presented on Chapter 5.

WorldDynamics.World3.NonRenewable.fig_26 — Method

Reproduce Fig 5.26. The original figure is presented on Chapter 5.

WorldDynamics.World3.NonRenewable.fig_28 — Method

Reproduce Fig 5.28. The original figure is presented on Chapter 5.

WorldDynamics.World3.NonRenewable.fig_29 — Method

Reproduce Fig 5.29. The original figure is presented on Chapter 5.

WorldDynamics.World3.NonRenewable.fig_30 — Method

Reproduce Fig 5.30. The original figure is presented on Chapter 5.

One level population system

WorldDynamics.World3.Pop1.fig_100 — Method

Reproduce Fig 2.100. The original figure is presented on Chapter 2.

WorldDynamics.World3.Pop1.fig_103a — Method

Reproduce Fig 2.103a. The original figure is presented on Chapter 2.

WorldDynamics.World3.Pop1.fig_103b — Method

Reproduce Fig 2.103b. The original figure is presented on Chapter 2.

WorldDynamics.World3.Pop1.fig_103c — Method

Reproduce Fig 2.103c. The original figure is presented on Chapter 2.

WorldDynamics.World3.Pop1.fig_84 — Method

Reproduce Fig 2.84. The original figure is presented on Chapter 2.

WorldDynamics.World3.Pop1.fig_85 — Method

Reproduce Fig 2.85. The original figure is presented on Chapter 2.

WorldDynamics.World3.Pop1.fig_86 — Method

Reproduce Fig 2.86. The original figure is presented on Chapter 2.

WorldDynamics.World3.Pop1.fig_87 — Method

Reproduce Fig 2.87. The original figure is presented on Chapter 2.

WorldDynamics.World3.Pop1.fig_88 — Method

Reproduce Fig 2.88. The original figure is presented on Chapter 2.

WorldDynamics.World3.Pop1.fig_89 — Method

Reproduce Fig 2.89. The original figure is presented on Chapter 2.

WorldDynamics.World3.Pop1.fig_90 — Method

Reproduce Fig 2.90. The original figure is presented on Chapter 2.

WorldDynamics.World3.Pop1.fig_91 — Method

Reproduce Fig 2.91. The original figure is presented on Chapter 2.

WorldDynamics.World3.Pop1.fig_93 — Method

Reproduce Fig 2.93. The original figure is presented on Chapter 2.

WorldDynamics.World3.Pop1.fig_96 — Method

Reproduce Fig 2.96. The original figure is presented on Chapter 2.

WorldDynamics.World3.Pop1.fig_97 — Method

Reproduce Fig 2.97. The original figure is presented on Chapter 2.

WorldDynamics.World3.Pop1.fig_98 — Method

Reproduce Fig 2.98. The original figure is presented on Chapter 2.

WorldDynamics.World3.Pop1.fig_99 — Method

Reproduce Fig 2.99. The original figure is presented on Chapter 2.

Four level population system

WorldDynamics.World3.Pop4.fig_100 — Method

Reproduce Fig 2.100. The original figure is presented on Chapter 2.

WorldDynamics.World3.Pop4.fig_101a — Method

Reproduce Fig 2.101a. The original figure is presented on Chapter 2.

WorldDynamics.World3.Pop4.fig_101b — Method

Reproduce Fig 2.101b. The original figure is presented on Chapter 2.

WorldDynamics.World3.Pop4.fig_101c — Method

Reproduce Fig 2.101c. The original figure is presented on Chapter 2.

WorldDynamics.World3.Pop4.fig_102a — Method

Reproduce Fig 2.102a. The original figure is presented on Chapter 2.

WorldDynamics.World3.Pop4.fig_102b — Method

Reproduce Fig 2.102b. The original figure is presented on Chapter 2.

WorldDynamics.World3.Pop4.fig_102c — Method

Reproduce Fig 2.102c. The original figure is presented on Chapter 2.

WorldDynamics.World3.Pop4.fig_104a — Method

Reproduce Fig 2.104a. The original figure is presented on Chapter 2.

WorldDynamics.World3.Pop4.fig_104b — Method

Reproduce Fig 2.104b. The original figure is presented on Chapter 2.

WorldDynamics.World3.Pop4.fig_104c — Method

Reproduce Fig 2.104c. The original figure is presented on Chapter 2.

WorldDynamics.World3.Pop4.fig_105a — Method

Reproduce Fig 2.105a. The original figure is presented on Chapter 2.

WorldDynamics.World3.Pop4.fig_105b — Method

Reproduce Fig 2.105b. The original figure is presented on Chapter 2.

WorldDynamics.World3.Pop4.fig_105c — Method

Reproduce Fig 2.105c. The original figure is presented on Chapter 2.

WorldDynamics.World3.Pop4.fig_106a — Method

Reproduce Fig 2.106a. The original figure is presented on Chapter 2.

WorldDynamics.World3.Pop4.fig_106b — Method

Reproduce Fig 2.106b. The original figure is presented on Chapter 2.

WorldDynamics.World3.Pop4.fig_106c — Method

Reproduce Fig 2.106c. The original figure is presented on Chapter 2.

WorldDynamics.World3.Pop4.fig_107a — Method

Reproduce Fig 2.107a. The original figure is presented on Chapter 2.

WorldDynamics.World3.Pop4.fig_107b — Method

Reproduce Fig 2.107b. The original figure is presented on Chapter 2.

WorldDynamics.World3.Pop4.fig_107c — Method

Reproduce Fig 2.107c. The original figure is presented on Chapter 2.

WorldDynamics.World3.Pop4.fig_108a — Method

Reproduce Fig 2.108a. The original figure is presented on Chapter 2.

WorldDynamics.World3.Pop4.fig_108b — Method

Reproduce Fig 2.108b. The original figure is presented on Chapter 2.

WorldDynamics.World3.Pop4.fig_108c — Method

Reproduce Fig 2.108c. The original figure is presented on Chapter 2.

WorldDynamics.World3.Pop4.fig_109a — Method

Reproduce Fig 2.109a. The original figure is presented on Chapter 2.

WorldDynamics.World3.Pop4.fig_109b — Method

Reproduce Fig 2.109b. The original figure is presented on Chapter 2.

WorldDynamics.World3.Pop4.fig_109c — Method

Reproduce Fig 2.109c. The original figure is presented on Chapter 2.

WorldDynamics.World3.Pop4.fig_110a — Method

Reproduce Fig 2.110a. The original figure is presented on Chapter 2.

WorldDynamics.World3.Pop4.fig_110b — Method

Reproduce Fig 2.110b. The original figure is presented on Chapter 2.

WorldDynamics.World3.Pop4.fig_110c — Method

Reproduce Fig 2.110c. The original figure is presented on Chapter 2.

WorldDynamics.World3.Pop4.fig_111a — Method

Reproduce Fig 2.111a. The original figure is presented on Chapter 2.

WorldDynamics.World3.Pop4.fig_111b — Method

Reproduce Fig 2.111b. The original figure is presented on Chapter 2.

WorldDynamics.World3.Pop4.fig_111c — Method

Reproduce Fig 2.111c. The original figure is presented on Chapter 2.

WorldDynamics.World3.Pop4.fig_112a — Method

Reproduce Fig 2.112a. The original figure is presented on Chapter 2.

WorldDynamics.World3.Pop4.fig_112b — Method

Reproduce Fig 2.112b. The original figure is presented on Chapter 2.

WorldDynamics.World3.Pop4.fig_112c — Method

Reproduce Fig 2.112c. The original figure is presented on Chapter 2.

WorldDynamics.World3.Pop4.fig_113a — Method

Reproduce Fig 2.113a. The original figure is presented on Chapter 2.

WorldDynamics.World3.Pop4.fig_113b — Method

Reproduce Fig 2.113b. The original figure is presented on Chapter 2.

WorldDynamics.World3.Pop4.fig_113c — Method

Reproduce Fig 2.113c. The original figure is presented on Chapter 2.

WorldDynamics.World3.Pop4.fig_84 — Method

Reproduce Fig 2.84. The original figure is presented on Chapter 2.

WorldDynamics.World3.Pop4.fig_85 — Method

Reproduce Fig 2.85. The original figure is presented on Chapter 2.

WorldDynamics.World3.Pop4.fig_86 — Method

Reproduce Fig 2.86. The original figure is presented on Chapter 2.

WorldDynamics.World3.Pop4.fig_87 — Method

Reproduce Fig 2.87. The original figure is presented on Chapter 2.

WorldDynamics.World3.Pop4.fig_88 — Method

Reproduce Fig 2.88. The original figure is presented on Chapter 2.

WorldDynamics.World3.Pop4.fig_89 — Method

Reproduce Fig 2.89. The original figure is presented on Chapter 2.

WorldDynamics.World3.Pop4.fig_90 — Method

Reproduce Fig 2.90. The original figure is presented on Chapter 2.

WorldDynamics.World3.Pop4.fig_91 — Method

Reproduce Fig 2.91. The original figure is presented on Chapter 2.

WorldDynamics.World3.Pop4.fig_93 — Method

Reproduce Fig 2.93. The original figure is presented on Chapter 2.

WorldDynamics.World3.Pop4.fig_96 — Method

Reproduce Fig 2.96. The original figure is presented on Chapter 2.

WorldDynamics.World3.Pop4.fig_97 — Method

Reproduce Fig 2.97. The original figure is presented on Chapter 2.

WorldDynamics.World3.Pop4.fig_98 — Method

Reproduce Fig 2.98. The original figure is presented on Chapter 2.

WorldDynamics.World3.Pop4.fig_99 — Method

Reproduce Fig 2.99. The original figure is presented on Chapter 2.

Fifteen level population system

WorldDynamics.World3.Pop15.fig_100 — Method

Reproduce Fig 2.100. The original figure is presented on Chapter 2.

WorldDynamics.World3.Pop15.fig_84 — Method

Reproduce Fig 2.84. The original figure is presented on Chapter 2.

WorldDynamics.World3.Pop15.fig_85 — Method

Reproduce Fig 2.85. The original figure is presented on Chapter 2.

WorldDynamics.World3.Pop15.fig_86 — Method

Reproduce Fig 2.86. The original figure is presented on Chapter 2.

WorldDynamics.World3.Pop15.fig_87 — Method

Reproduce Fig 2.87. The original figure is presented on Chapter 2.

WorldDynamics.World3.Pop15.fig_88 — Method

Reproduce Fig 2.88. The original figure is presented on Chapter 2.

WorldDynamics.World3.Pop15.fig_89 — Method

Reproduce Fig 2.89. The original figure is presented on Chapter 2.

WorldDynamics.World3.Pop15.fig_90 — Method

Reproduce Fig 2.90. The original figure is presented on Chapter 2.

WorldDynamics.World3.Pop15.fig_91 — Method

Reproduce Fig 2.91. The original figure is presented on Chapter 2.

WorldDynamics.World3.Pop15.fig_93 — Method

Reproduce Fig 2.93. The original figure is presented on Chapter 2.

WorldDynamics.World3.Pop15.fig_94a — Method

Reproduce Fig 2.94a. The original figure is presented on Chapter 2.

WorldDynamics.World3.Pop15.fig_94b — Method

Reproduce Fig 2.94b. The original figure is presented on Chapter 2.

WorldDynamics.World3.Pop15.fig_94c — Method

Reproduce Fig 2.94c. The original figure is presented on Chapter 2.

WorldDynamics.World3.Pop15.fig_96 — Method

Reproduce Fig 2.96. The original figure is presented on Chapter 2.

WorldDynamics.World3.Pop15.fig_97 — Method

Reproduce Fig 2.97. The original figure is presented on Chapter 2.

WorldDynamics.World3.Pop15.fig_98 — Method

Reproduce Fig 2.98. The original figure is presented on Chapter 2.

WorldDynamics.World3.Pop15.fig_99 — Method

Reproduce Fig 2.99. The original figure is presented on Chapter 2.

World3 system

WorldDynamics.World3.fig_10 — Method

Reproduce Fig 7.10. The original figure is presented on Chapter 7.

WorldDynamics.World3.fig_11 — Method

Reproduce Fig 7.11. The original figure is presented on Chapter 7.

WorldDynamics.World3.fig_13 — Method

Reproduce Fig 7.13. The original figure is presented on Chapter 7.

WorldDynamics.World3.fig_14 — Method

Reproduce Fig 7.14. The original figure is presented on Chapter 7.

WorldDynamics.World3.fig_15 — Method

Reproduce Fig 7.15. The original figure is presented on Chapter 7.

WorldDynamics.World3.fig_16 — Method

Reproduce Fig 7.16. The original figure is presented on Chapter 7.

WorldDynamics.World3.fig_18 — Method

Reproduce Fig 7.18. The original figure is presented on Chapter 7.

WorldDynamics.World3.fig_19 — Method

Reproduce Fig 7.19. The original figure is presented on Chapter 7.

WorldDynamics.World3.fig_2 — Method

Reproduce Fig 7.2. The original figure is presented on Chapter 7.

WorldDynamics.World3.fig_20 — Method

Reproduce Fig 7.20. The original figure is presented on Chapter 7.

WorldDynamics.World3.fig_21 — Method

Reproduce Fig 7.21. The original figure is presented on Chapter 7.

WorldDynamics.World3.fig_22 — Method

Reproduce Fig 7.22. The original figure is presented on Chapter 7.

WorldDynamics.World3.fig_23 — Method

Reproduce Fig 7.23. The original figure is presented on Chapter 7.

WorldDynamics.World3.fig_24 — Method

Reproduce Fig 7.24. The original figure is presented on Chapter 7.

WorldDynamics.World3.fig_26 — Method

Reproduce Fig 7.26. The original figure is presented on Chapter 7.

WorldDynamics.World3.fig_27 — Method

Reproduce Fig 7.27. The original figure is presented on Chapter 7.

WorldDynamics.World3.fig_3 — Method

Reproduce Fig 7.3. The original figure is presented on Chapter 7.

WorldDynamics.World3.fig_30 — Method

Reproduce Fig 7.30. The original figure is presented on Chapter 7.

WorldDynamics.World3.fig_32 — Method

Reproduce Fig 7.32. The original figure is presented on Chapter 7.

WorldDynamics.World3.fig_34 — Method

Reproduce Fig 7.34. The original figure is presented on Chapter 7.

WorldDynamics.World3.fig_35 — Method

Reproduce Fig 7.35. The original figure is presented on Chapter 7.

WorldDynamics.World3.fig_36 — Method

Reproduce Fig 7.36. The original figure is presented on Chapter 7.

WorldDynamics.World3.fig_37 — Method

Reproduce Fig 7.37. The original figure is presented on Chapter 7.

WorldDynamics.World3.fig_38 — Method

Reproduce Fig 7.38. The original figure is presented on Chapter 7.

WorldDynamics.World3.fig_39 — Method

Reproduce Fig 7.39. The original figure is presented on Chapter 7.

WorldDynamics.World3.fig_4 — Method

Reproduce Fig 7.4. The original figure is presented on Chapter 7.

WorldDynamics.World3.fig_41 — Method

Reproduce Fig 7.41. The original figure is presented on Chapter 7.

WorldDynamics.World3.fig_5 — Method

Reproduce Fig 7.5. The original figure is presented on Chapter 7.

WorldDynamics.World3.fig_6 — Method

Reproduce Fig 7.6. The original figure is presented on Chapter 7.

WorldDynamics.World3.fig_7 — Method

Reproduce Fig 7.7. The original figure is presented on Chapter 7.

WorldDynamics.World3.fig_8 — Method

Reproduce Fig 7.8. The original figure is presented on Chapter 7.

WorldDynamics.World3.fig_9 — Method

Reproduce Fig 7.9. The original figure is presented on Chapter 7.